chore: :building_construction: upgrade vitepress template
this PR updates vitepress from 1.0.2 to 1.6.3. This was achieved by recreating the template and then copying the markdown files
