--- .vitepress/theme/index.css
@@ -1,7 +1,0 @@
-.medium-zoom-overlay {
-    z-index: 30;
-  }
-  
-  .medium-zoom-image {
-    z-index: 31;
-  }
