+++ package.json
@@ -10,7 +10,7 @@
   "keywords": [],
   "license": "UNLICENSED",
   "devDependencies": {
-    "vitepress": "^1.0.2"
+    "vitepress": "^1.6.3"
   },
   "dependencies": {
     "medium-zoom": "^1.1.0"
